Give the repository the gates its own philosophy demands
The stated principle here is "a gate that is skipped is not a gate" — and the
toolchain didn't meet it: no linter, no type checker despite shipping
`py.typed`, no dependency automation despite byte-identity resting on renderer
libraries' clock and zip behaviour, slow tests that ran nowhere, and no human
contributor surface beside an exemplary agent one. This closes the gap, and
each gate gates.

**Ruff: 4,275 findings → 0, blocking in CI.** ~283 autofixed, ~65 hand-fixed,
six rules disabled with the house pattern each protects named at the config
(E501 guards long-form why-comment prose, measured p99=96; E402 guards the
registration-at-import seam; RUF022 guards semantically-grouped `__all__`
sections). What the linter caught was real: dead provenance in
`facets.resolve` written and never read, seven `pytest.raises(match=)`
patterns with unescaped regex metachars that could pass on wrong messages, 22
mutable `[]` defaults in generator closures, 16 noqa directives suppressing
nothing. Byte-identity after ~150 files of fixes proven twice — the default
corpus against a `git archive HEAD` baseline, then `tools/sweep.py --mode
archive` across four sampled configurations, 4/4 identical.

**Mypy: measured 520 errors at strict, 254 at default — gated at the honest
scope, blocking, exit 0.** Full `src/worldloom` at default strictness with a
named 58-module `ignore_errors` debt list: 104 of 163 modules held to the
standard today, the thin waist among them, new modules blocking by default,
and the list allowed only to shrink. The checker paid for itself on day one:
`detail.py` had declared its own duplicate `Model` base — making
`DetailTable` a different Model from the one `Collection` is bounded by — and
`Collection.where/filter/sort_by` erased subclass types, fixed with `Self`,
which deleted an in-tree workaround.

**The slow suite was red, and had been red unwitnessed.** `pytest -m slow` is
deselected by `addopts` and ran nowhere; its first real execution found
`test_low_density_shrinks_the_optional_fan_out_without_losing_questions`
failing at HEAD — the approval-provenance evaluation family postdates the
test, mints questions on a document low density trims, and the blanket
set-equality broke. That is the repository's own argument about skipped gates,
demonstrated on its own suite. The fix tightens rather than loosens: every
lost question must require a trimmed document, and losing a question whose
documents survive is still a failure. A weekly cron job now runs them, born
green.

Also: dependabot (renderer and data groups, with the byte-identity rationale
in the file), README badges for the three targets that exist — the PyPI badge
is deliberately absent with a dated comment, because the name 404s and a
broken badge is worse than none; CONTRIBUTING/SECURITY/issue templates that
ask for seed + recipe + version, since reproducibility is the product; a
minimal PR template; pre-commit with ruff and the docs drift check —
ruff-format refused with the measurement stated (it would reflow 294 of 309
files of deliberate comment paragraphing); coverage on one matrix leg,
report-only by design and saying why in the workflow comment (a threshold is
a proxy gate that invites test theatre) — measured 90%; CI installs on uv,
verified locally first, except the package job which deliberately keeps pip
because it proves the wheel under the installer users actually run; and a
reduced macOS leg on the nightly sweep, shipped with the honest caveat in its
comment: the OS-independence and byte-identity claims have only ever been
proven on ubuntu, this leg cannot be verified from this container, and a red
run is a finding about the claim rather than noise.

Both new gates are pinned exactly (ruff 0.15.8, mypy 2.3.1): a tool release
that enables new rules must not turn an unrelated PR red; bumping the pin is
a deliberate PR that fixes what the newer tool finds.

Verified: full suite 3,005 passed; slow suite 3/3; `validate retail-close`
1,283 unchanged; `docs --check` current; ruff and mypy exit 0 under CI's
exact pins in the project environment; all five YAMLs parse; default corpus
byte-identical to the HEAD baseline.

+++ src/worldloom/episodes.py
@@ -28,7 +28,8 @@ from __future__ import annotations
 
 import json
 from collections.abc import Iterable, Sequence
-from dataclasses import dataclass, field as _dataclass_field
+from dataclasses import dataclass
+from dataclasses import field as _dataclass_field
 from datetime import datetime
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal
@@ -39,8 +40,8 @@ from . import benchmark as benchmark_module
 from . import detail as detail_module
 from . import validate as validate_module
 from .benchmark import EvalSpec, QuestionFamily  # re-exported: a spec's own vocabulary
-from .models import Authority, CanonicalFact, EnterpriseEvent, Quantity
 from .models import ArtifactIntent as ArtifactIntentModel
+from .models import Authority, CanonicalFact, EnterpriseEvent, Quantity
 
 if TYPE_CHECKING:  # pragma: no cover
     from .ids import Minter
@@ -1424,7 +1425,7 @@ def run(
             if head == "pct_of":
                 pct = primitives.level(physics, fk.parameter, rng.derive(f"kind/{kind}"))
                 values[kind] = shaped(
-                    kind, operands, lambda a: int(round(a * float(pct) / 100)),
+                    kind, operands, lambda a: round(a * float(pct) / 100),
                 )
             elif head == "ratio_pct":
                 a, b = operands
@@ -1472,7 +1473,7 @@ def run(
                 values[kind] = shaped(kind, operands, lambda a, b: round(a - b, 2))
             elif head == "units_of":
                 values[kind] = shaped(
-                    kind, operands, lambda v, r: int(round(v * 1000 / r)),
+                    kind, operands, lambda v, r: round(v * 1000 / r),
                 )
             elif head == "prior":
                 # Zero when no prior period holds the slot — minted, not
@@ -1507,7 +1508,7 @@ def run(
                 # rediscovered: it is arbitrary between cohorts of equal
                 # weight, and the only thing that matters is that it is the
                 # same arbitrary choice every rebuild makes.
-                allocated = primitives.rollup(int(round(total)), weights)
+                allocated = primitives.rollup(round(total), weights)
                 values[kind] = ("cohort", dict(zip(grid, allocated)))
             elif head == "prior_in_cohort":
                 # The diagonal step. `prior(K)` cannot express this: it walks
@@ -2201,7 +2202,7 @@ class AuthoredEpisode:
 
 # The recipe verb, registered from this module exactly as each vertical's
 # scenario registers its own — an authored episode's replay costs core nothing.
-from . import recipe as _recipe  # noqa: E402
+from . import recipe as _recipe
 
 _recipe.register_step("AuthoredEpisode", ("episode", "period"), AuthoredEpisode)
